Update DbUrlList to use symbol operators
The DbUrlList.getNextUrl function was failing on Postgres when I tried
to use it. After some exploration, I discovered the use of `$lt`. I
changed the code to use [Op.lt] and getNextUrl started working as
expected.

Also, Sequelize says it is highly advised to use symbol operators from
Sequelize.Op and not depend on any string based operators.